feat: 1.启用 redis 缓存 2.token 鉴权使用 redis 缓存

+++ dice-server/src/main/java/com/bihell/dice/security/AuthToken.java
@@ -22,7 +22,7 @@ public class AuthToken {
      * Expired in. (seconds)
      */
     @JsonProperty("expired_in")
-    private int expiredIn;
+    private long expiredIn;
 
     /**
      * Refresh token.
